bugfix and change in folder structure
