+++ package.json
@@ -19,7 +19,6 @@
     "formik": "^2.0.8",
     "formik-material-ui": "^1.0.0",
     "prop-types": "^15.7.2",
-    "pusher-js": "^5.0.3",
     "react": "^16.12.0",
     "react-calendar": "^2.19.2",
     "react-dom": "^16.12.0",
